requirements.txt
pynetbox==7.3.3
python-dotenv==1.0.1
icmplib==3.0.4
influxdb-client==1.43.0
requests==2.32.3
